Modify entities, change id from type int to Long. Add AttractionDto. Add margin to login component.

+++ wserver/src/main/java/com/wanderdrop/wserver/model/Attraction.java
@@ -17,7 +17,7 @@ public class Attraction {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "attraction_id")
-    private int attractionId;
+    private Long attractionId;
 
     @Column(name = "attraction_name", nullable = false)
     private String name;
