+++ wserver/src/main/java/com/wanderdrop/wserver/model/DeletionReason.java
@@ -11,7 +11,7 @@ public class DeletionReason {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name=" deletion_reason_id")
-    private int deletionReasonId;
+    private Long deletionReasonId;
 
     @Column(name= "reason_message", nullable = false)
     private String reasonMessage;
